pyproject.toml
[tool.poetry]
name = "konan-titanic"
version = "0.1.0"
description = "A Konan-ready ML model, packaged and distributed as a Docker image, to predict the survival of a Titanic passenger"
authors = ["Synapse Analytics"]
readme = "README.md"
repository = "https://github.com/SynapseAnalytics/konan-titanic-model"

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^1.4.4"
sklearn = "^0.0"
joblib = "^1.1.0"
PyYAML = "^6.0"
konan-sdk = "^1.3.0"

[tool.poetry.dev-dependencies]
flake8 = "^5.0.4"

[tool.semantic_release]
branch = "main"
major_on_zero = false
version_variable = [
    "pyproject.toml:version",
]
version_toml = [
    "pyproject.toml:tool.poetry.version"
]
version_pattern = [
    ".konan.example:KONAN_APP_VERSION=v{version}",
]
commit_author = "github-actions <actions@github.com>"
upload_to_repository = false
upload_to_release = true
build_command = false

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
